Add a 'Provided' type to match on 'Required'/'NotRequired' values.
I learnt (little bughunt involved) that once there is an easy way
to switch the type of combinator from `NotRequired` to `Required`,
it's very prone to do that and forget to make corresponding update
to the handler.

If the optional parameters are passed as 'Maybe a' and required as
'a', such mistake won't be caught, as 'Maybe' happen to have instances
for quite many things (incluiding `FromJSON`, `FromHttpApiData`, etc.),
so if one tries to decode a required parameter of type `Maybe a`
everything that doesn't look like "Just ..." would result into a `400`.

Here I add a type 'Provided' that has maybe-like constructors, and a
phantom 'Required' type variable. GADT is used to enforce that the
`NotProvided` value can be only `NotRequried`.

Using it for all the different paramers would break a lot of compatibility,
so we probably could start by applying it only to the newly added
optional combinators (like `OptionalReqBody` from the previous commit).

I am not yet sure how feasible that is, but it seems like a nice
design, so I'll add it to get some comments.