Add HasServer instance for (ReqBody' 'NotRequired)
The `route` implementation checks request headers (see code for
details) and if they don't indicate that there is a body, gives up
and returns `NotProvided`.

An alternative behaviour is to always try decoding the body, but
return 'NotProvided' if that fails (for instance due to the body
being obviously empty).